theme/templates/index.html: force conversion of category to string
Pelican started to mis-render the page when running it locally.

The exact reason is still unclear, but this was apparently caused by "news"
in category[0] not being considered a string (though, when output, it would
be indistinguishable from one), causing the comparisons to always fail.

See also
http://stackoverflow.com/questions/12623768/compare-two-variables-in-jinja2-template

This fix or work-around should produce no visible difference on neo900.org